requirements.txt
apturl==0.5.2
arrow==0.17.0
astroid==2.4.2
attrs==19.3.0
basicauth==0.4.1
bcrypt==3.1.7
black==20.8b1
blinker==1.4
boto==2.49.0
boto3==1.15.16
Brlapi==0.7.0
cached-property==1.5.1
chrome-gnome-shell==0.0.0
colorama==0.4.3
command-not-found==0.3
croniter==0.3.34
cryptography==3.1
cupshelpers==1.0
db-fix==0.1
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
distro-info==0.23ubuntu1
dj-database-url==0.5.0
dj-static==0.0.6
django-cors-headers==3.7.0
django-cron==0.5.1
django-crontab==0.7.1
django-environ==0.4.5
django-filter==2.2.0
django-mathfilters==1.0.0
django-otp==0.8.1
djangorestframework==3.12.4
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
gitignore==0.0.8
gunicorn==19.9.0
haversine==2.1.2
importlib-metadata==1.5.0
isort==5.5.4
jsonschema==3.2.0
language-selector==0.1
launchpadlib==1.10.13
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
mccabe==0.6.1
messenger==0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
olefile==0.46
pandas==1.2.4
paramiko==2.6.0
pep8==1.7.1
pexpect==4.6.0
Pillow==7.2.0
pre-commit==2.13.0
protobuf==3.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycryptodome==3.9.9
pycups==1.9.73
PyGObject==3.36.0
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pyotp==2.3.0
PyPDF2==1.26.0
PyPDF4==1.27.0
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.5
python-debian==0.1.36ubuntu1
pyxdg==0.26
redis==3.4.1
reportlab==3.5.34
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
schedule==0.6.0
simplejson==3.16.0
solution==0.4
systemd-python==234
texttable==1.6.2
twilio==6.46.0
ubuntu-advantage-tools==27.0
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
usb-creator==0.3.7
websocket-client==0.53.0
whitenoise==4.1.2
xkit==0.0.0
xl2dict==0.1.5
zipp==1.0.0
